Add files via upload
Hey Rob, I did a very quick Sketchup of a compact unit - sized to a 15inch laptop screen, depth to accomodate a raspberry pi, speakers, cabling, controller supports and so on, and the lid designed to accomodate the controllers without need for removal on closure. In terms of design / construction, this would be a very simple build and easy to make fairly elegant / add decorative elements, lighting etc etc etc. Anyway, thought it might be food for thought vs the large cabinet design. Happy to do other designs once we have a few more confirmed elements which i can build to scale in Sketchup.